hydra-optuna/requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
optuna==2.10.1
scikit_learn==1.4.2
